+++ verik-compiler/src/main/kotlin/io/verik/compiler/main/StageType.kt
@@ -25,6 +25,7 @@ enum class StageType {
     MID_CHECK,
     SPECIALIZE,
     RESOLVE,
+    EVALUATE,
     INTERPRET,
     UPPER_TRANSFORM,
     LOWER_TRANSFORM,